+++ src/aliases.ts
@@ -1,7 +1,6 @@
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type ANY = any;
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type json = any;
-
 export type ToStringable = {toString: () => string};
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export type ANY = any;
+export type json = any;
+/* eslint-enable @typescript-eslint/no-explicit-any */
