+++ src/fate/code.ts
@@ -1,4 +1,4 @@
-export enum AckCode {
+export enum FateCode {
 	NOT_SET = 0,
 	SUCCESS = 1,
 	FAILURE = -1
