+++ src/fate/options.ts
@@ -1,4 +1,4 @@
-export interface AckOptions<T = unknown> {
+export interface FateOptions<T = unknown> {
 	errorThreshold?: number;
 	payload?: T;
 	serialized?: string;
