+++ src/index.ts
@@ -1,4 +1,4 @@
-export {Ack} from './ack';
-export {AckCode} from './ack/code';
-export {AckOptions} from './ack/options';
-export {AckState} from './ack/state';
+export {Fate} from './fate';
+export {FateCode} from './fate/code';
+export {FateOptions} from './fate/options';
+export {FateState} from './fate/state';
